Move scale- and svg- related helpers to the /helpers directory
