+++ src/features/graph-view-2/hooks/useZoom.ts
@@ -1,4 +1,6 @@
 import { RefObject, useEffect, useState } from "react";
+import { scaleCenterToMatchTarget } from "../helpers/scale";
+import { getPointInSvgSpace } from "../helpers/svg";
 
 export type Position = [x: number, y: number];
 export type Offset = [x: number, y: number];
@@ -18,25 +20,15 @@ export const useZoom = (
       return;
     }
 
-    // prettier-ignore
     const mouseWheelHandler = (event: WheelEvent) => {
       const scale = getScaleFactor(-event.deltaY);
       setZoom(zoom * scale);
 
       const mousePosition: Position = [event.clientX, event.clientY];
+      // prettier-ignore
       const mousePositionInSVGSpace = getPointInSvgSpace(mousePosition, svgElement);
 
-      const scaleCenterToMatchTarget = (center: Position, target: Position, scale: number): Position => {
-        const newCenterX = scaleAxisCenterToMatchTarget(center[0], target[0], scale);
-        const newCenterY = scaleAxisCenterToMatchTarget(center[1], target[1], scale);
-        return [newCenterX, newCenterY];
-      }
-
-      const scaleAxisCenterToMatchTarget = (center: number, target: number, scale: number): number => {
-        const delta = target - center;
-        return target - delta / scale;
-      }
-
+      // prettier-ignore
       const newCenter = scaleCenterToMatchTarget(center, mousePositionInSVGSpace, scale);
       setCenter(newCenter);
     };
@@ -57,13 +49,4 @@ const getScaleFactor = (delta: number): number => {
     return 1 + scale;
   }
   return 1 + delta / 10 / Math.abs(delta || 1);
-};
-
-const getPointInSvgSpace = ([x, y]: Position, svg: SVGSVGElement): Position => {
-  const point = new DOMPoint(x, y);
-  const svgViewportOffsetMatrix = svg.getScreenCTM()?.inverse();
-  const { x: pointInSVGSpaceX, y: pointInSVGSpaceY } = point.matrixTransform(
-    svgViewportOffsetMatrix
-  );
-  return [pointInSVGSpaceX, pointInSVGSpaceY];
 };
